+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * `Repos`:
     * `Cache`:
         * New transformer type: `ReadCRUDFromKeyValueRepo`
+        * New transformer type: `ReadKeyValueFromCRUDRepo`
+* `Pagination`:
+    * New `paginate` extensions with `reversed` support for `List`/`Set`
 
 ## 0.16.9